Use a custom find-files procedure
CHICKEN core's find-files uses `glob' for listing files in
directories.  `glob' uses regex behind the scenes and is significantly
slower than `directory' for when we want to list all files.  The
variant of `find-files' introduced by this commit is a version fo
CHICKEN core's `find-files' using `directory' instead of `glob'.