Ignore SIGPIPE
Ignore SIGPIPE, otherwise with-output-to-pipe would make fu exit
without displaying the prompt, in case user terminates the pager.
That may happen on a large number of options going throu the pipe, and
user finds what he/she is looking for at the first options, then
he/she it terminates the pager.