Restore support for git grep (handle submodules)
Use "git grep" in case "fu grep" is called from a git repository which
is not a submodule.

In case of using "fu grep" in the root of a project with submodules,
resort to regular grep (i.e., not "git grep" -- either grep or rg), as
"git submodule grep" doesn't exist and implementing it is not trivial.

Bump version to 1.1.0.